+++ testdata/ExecDirect_test.go
@@ -7,3 +7,15 @@ func TestExecDirect(t *testing.T) {
 		t.Error("Error in ExecDirect")
 	}
 }
+
+//ExecDirect will execute the query without prepare
+func ExecDirect() error {
+        db := Createconnection()
+        defer db.Close()
+        _, err := db.Query("select * from rocket")
+        if err != nil {
+                return err
+        }
+        return nil
+}
+
