+++ testdata/decimal_test.go
@@ -1,6 +1,8 @@
 package main
 
 import (
+        "fmt"
+        "strings"
         "testing"
 )
 
@@ -10,3 +12,46 @@ func TestDecimalColumn(t *testing.T) {
 		t.Error("Error at DecimalColumn")
 	}
 }
+
+func DecimalColumn() int {
+        var tableOne string= "godecmaltable"
+
+        db := Createconnection()
+
+        db.Query("DROP table " + tableOne)
+
+        _, err := db.Query("CREATE table " + tableOne + "(col1 DECIMAL(30, 2))")
+        if !strings.Contains(fmt.Sprint(err), "did not create a result set") {
+                return 0
+        }
+
+        _, err = db.Query("INSERT into " + tableOne + "(col1) values(9999999999999999999999999999.99)")
+        if !strings.Contains(fmt.Sprint(err), "did not create a result set") {
+                 return 0
+        }
+
+        _, err = db.Query("INSERT into " + tableOne + "(col1) values(99999999999999999999)")
+        if !strings.Contains(fmt.Sprint(err), "did not create a result set") {
+                 return 0
+        }
+
+        rows, err2 := db.Query("SELECT * from " + tableOne)
+        if err2 != nil {
+                return 0
+        }
+
+        defer rows.Close()
+        for rows.Next() {
+              var f  string
+              err = rows.Scan(&f)
+              if err != nil {
+                return 0
+              }
+              //fmt.Printf("%v \n", f)
+        }
+
+        db.Query("DROP table " + tableOne)
+
+        return 1
+}
+
