+++ testdata/querycreatedroptable_test.go
@@ -9,3 +9,23 @@ func TestQueryCreateDropTable(t *testing.T) {
                 t.Error("Error at CreateDropTable")
         }
 }
+
+func QueryCreateDropTable() int {
+	db := Createconnection()
+        if db == nil {
+                return 0
+        }
+
+        err := QueryCreateTable(db)
+        if err != nil {
+                return 0
+        }
+
+        err = QueryDropTable(db)
+        if err != nil {
+                return 0
+        }
+
+        return 1
+}
+
